vision-server/requirements.txt
ultralytics
fastapi
uvicorn
python-multipart
httpx
pillow
